Allows filereader.py to extract SpecificCharacterSet from Datasets
During the reading of a DICOM file, most elements are read in as RawDataElements by the data_element_generator. The one exception is a SQ (a list of Datasets). When a SQ is encountered, a new Dataset is created, therefore the RawDataElements of that Dataset are converted to DataElements. In python3, these DataElements that contain string VRs should be decoded from bytes to a string literal. It is necessary for Datasets inside of a SQ to know the parent Dataset's encoding so that they can use the as their default if (0008,0005) isn't found within itself.

This commit simply passes around the encoding that is found in the Dataset. Currently this encoding isn't _actually_ applied to the values themselves