Corrects improper Dataset decoding of Datasets containing SQ elements with their own encodings provided
If a Dataset contains a SQ (a list of Datasets), it is possible that the contents of the SQ actually specify their own encoding ( Tag: (0008,0005)). If they don't, then they use the parent's encoding.

In the old way, the parents encoding was forced on ALL DataElements recursively (even those inside of other Datasets). This is because the Dataset.walk() method was used and it returned all DataElements.

Dataset.walk has been modified to accept an additional argument to define whether it is recursive or not. Dataset.decode() no longer uses the recursive and simply calls the decode() method of any Dataset in a Sequence it encounters.