Fixes python 3 bug in exception-checking unittests
The use of unittest.TestCase.failUnlessRaises() is deprecated and it is advisable to use
unittest.TestCase.assertRaises() instead. Previously, we had relied upon an overloaded
failUnlessRaises() method to provide the exception message for comparison; however,
when using the 2to3 conversion tool, it automatically replaced calls to failUnlessRaises()
with calls to assertRaises() thereby bypassing our overloaded method and resulting in
test failure.

In python 2.7+, it is possible to use a with statement for this assertion and we can determine
the exception message from the context manager. Currently this test uses the in_py3 check
due to the lack of this feature in 2.6. For all 2.x versions, the previous overloaded function is used.